Moving the hybrid s-t mesh definitions into their own sub-module. They still inherit from the Cartesian class.
